Events firebase setup explanation
Firebase firestore was set up so that you add an event to the "events" collection with any document id. Each document includes three fields; a "date" field for when the event expires, a "link" field for if you want the user to be directed to a URL or not, and a "title" field which is used for the foregrounds alt attribute and finding the images.

For Firebase storage, I had it set up so that the images are in the "events" folder, and the images should we named like so: Foreground: "(Title of event without spaces.)-fg.png", Background: "(Title of event without spaces.)-bg.png". An example would be: "WaterBaptism-bg.png".